requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
Django==1.9
django-crispy-forms==1.12.0
django-formtools==1.0
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
httplib2==0.9.2
idna==3.2
Pillow==8.3.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
